+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project follows [Semantic Versioning](https://semver.org/spec/v2.0.0.ht
 
 ## [Unreleased]
 
+### Added
+
+- Added MS Word option to the acceptance criteria report via a PDF/Word split button for single and multi-team tabs.
+
 ## [2.21.0] - 2026-09-09
 
 ### Added